Working progress... ES5 almost completed

+++ src/es5/statements/block_statement.js
@@ -1,7 +1,7 @@
 /* -.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-.
 * File Name   : block_statement.js
 * Created at  : 2017-08-18
-* Updated at  : 2019-03-02
+* Updated at  : 2019-03-30
 * Author      : jeefo
 * Purpose     :
 * Description :
@@ -14,10 +14,8 @@ _._._._._._._._._._._._._._._._._._._._._.*/
 
 // ignore:end
 
-const states_enum        = require("../enums/states_enum"),
-      precedence_enum    = require("../enums/precedence_enum"),
-      get_pre_comment    = require("../helpers/get_pre_comment"),
-      get_start_position = require("../helpers/get_start_position");
+const states_enum     = require("../enums/states_enum"),
+      precedence_enum = require("../enums/precedence_enum");
 
 module.exports = {
 	id         : "Block statement",
@@ -27,22 +25,23 @@ module.exports = {
     is : (token, parser) => {
         if (token.value === '{') {
             switch (parser.current_state) {
+                case states_enum.statement       :
                 case states_enum.block_statement :
-                    return true;
-                case states_enum.statement :
-                    // TODO: check make sure it's statement
                     return true;
             }
         }
 
         return false;
     },
+
 	initialize : (symbol, current_token, parser) => {
-        const statements  = [],
-              pre_comment = get_pre_comment(parser);
+        const statements   = [],
+              is_statement = parser.current_state === states_enum.statement;
+
+        parser.change_state("delimiter");
+        const open_curly_bracket = parser.next_symbol_definition.generate_new_symbol(parser);
 
         parser.prepare_next_state(null, true);
-
 		while (parser.next_token.value !== '}') {
             const statement = parser.get_next_symbol(precedence_enum.TERMINATION);
             if (! parser.is_terminated) {
@@ -57,11 +56,16 @@ module.exports = {
             parser.prepare_next_state(null, true);
 		}
 
-        symbol.pre_comment = pre_comment;
-        symbol.statements  = statements;
-        symbol.start       = get_start_position(pre_comment, current_token);
-        symbol.end         = parser.next_token.end;
+        const close_curly_bracket = parser.next_symbol_definition.generate_new_symbol(parser);
 
-        parser.terminate(symbol);
+        symbol.open_curly_bracket  = open_curly_bracket;
+        symbol.statements          = statements;
+        symbol.close_curly_bracket = close_curly_bracket;
+        symbol.start               = symbol.open_curly_bracket.start;
+        symbol.end                 = symbol.close_curly_bracket.end;
+
+        if (is_statement) {
+            parser.terminate(symbol);
+        }
     }
 };
