+++ src/es5/statements/expression_statement.js
@@ -1,7 +1,7 @@
 /* -.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-.-.
 * File Name   : expression_statement.js
 * Created at  : 2017-08-17
-* Updated at  : 2019-03-18
+* Updated at  : 2019-03-30
 * Author      : jeefo
 * Purpose     :
 * Description :
@@ -14,11 +14,23 @@ _._._._._._._._._._._._._._._._._._._._._.*/
 
 // ignore:end
 
-const states_enum                 = require("../enums/states_enum"),
-      precedence_enum             = require("../enums/precedence_enum"),
-      get_pre_comment             = require("../helpers/get_pre_comment"),
-      get_start_position          = require("../helpers/get_start_position"),
-      get_last_non_comment_symbol = require("../helpers/get_last_non_comment_symbol");
+const states_enum     = require("../enums/states_enum"),
+      precedence_enum = require("../enums/precedence_enum"),
+      get_right_value = require("../helpers/get_right_value");
+
+const expression_delimiters = ['[', '('];
+const expression_keywords = [
+    "new",
+    "void",
+    "delete",
+    "typeof",
+
+    "null",
+    "true",
+    "false",
+    "undefined",
+];
+const unary_expressions = [ "~", "!", "+", "-", "++", "--" ];
 
 module.exports = {
     id         : "Expression statement",
@@ -26,54 +38,63 @@ module.exports = {
     precedence : 40,
 
     is : (token, parser) => {
-        if (parser.current_state === states_enum.statement) {
-            switch (token.id) {
-                case "String" :
-                case "Number" :
-                    return true;
-                case "Identifier" :
-                    switch (token.value) {
-                        case "new" :
-                        case "void" :
-                        case "delete" :
-                        case "typeof" :
-
-                        case "null" :
-                        case "true" :
-                        case "false" :
-                        case "undefined" :
+        switch (parser.current_state) {
+            case states_enum.statement :
+                switch (token.id) {
+                    case "Slash"  :
+                    case "String" :
+                    case "Number" :
+                        return true;
+                    case "Identifier" :
+                        if (expression_keywords.includes(token.value)) {
                             return true;
-                    }
-                    return parser.symbol_table.reserved_words[token.value] === undefined;
-            }
+                        }
+                        return parser.symbol_table.reserved_words[token.value] === undefined;
+                    case "Delimiter" :
+                        return expression_delimiters.includes(token.value);
+                    case "Operator" :
+                        return unary_expressions.includes(token.value);
+                }
+                break;
         }
         return false;
     },
 
     initialize : (symbol, current_token, parser) => {
-        let post_comment  = null;
-        const pre_comment = get_pre_comment(parser);
+        let terminator = null;
+        const is_individual_block_statement = parser.current_state === states_enum.statement;
 
-        parser.current_symbol = null;
         parser.change_state("expression");
-        let expression = parser.get_next_symbol(precedence_enum.TERMINATION);
-        if (expression.id === "Comment") {
-            if (parser.next_token !== null && parser.next_token.value === ';') {
-                post_comment = expression;
-            }
 
-            expression = get_last_non_comment_symbol(parser);
+        // Labelled statement
+        if (parser.next_symbol_definition.id === "Identifier") {
+            parser.current_symbol   = parser.next_symbol_definition.generate_new_symbol(parser);
+            parser.previous_symbols = [parser.current_symbol];
+            parser.prepare_next_symbol_definition();
+
+            if (parser.next_token !== null && parser.next_token.value === ':') {
+                symbol.identifier = parser.current_symbol;
+                symbol.delimiter  = parser.next_symbol_definition.generate_new_symbol(parser);
+                return parser.change_state("labelled_statement");
+            }
         }
 
-        const asi = parser.next_token === null || parser.next_token.value !== ';';
+        parser.post_comment = null;
+        const expression = get_right_value(parser, precedence_enum.TERMINATION);
+        parser.current_symbol = parser.post_comment;
 
-        symbol.expression   = expression;
-        symbol.ASI          = asi;
-        symbol.pre_comment  = pre_comment;
-        symbol.post_comment = post_comment;
-        symbol.start        = get_start_position(pre_comment, expression);
-        symbol.end          = asi ? expression.end : parser.next_token.end;
+        if (parser.next_token !== null && parser.next_token.value === ';') {
+            parser.change_state("delimiter");
+            terminator = parser.next_symbol_definition.generate_new_symbol(parser);
+        }
 
-        parser.terminate(symbol);
+        symbol.expression = expression;
+        symbol.terminator = terminator;
+        symbol.start      = expression.start;
+        symbol.end        = terminator ? terminator.end : expression.end;
+
+        if (is_individual_block_statement) {
+            parser.terminate(symbol);
+        }
     }
 };
